Big rewrite of audit_given_nations()
It will now give no flag rather than a 0 when complying. I think this should be more aesthetic and easier to read if I am understanding it correctly. I also almost completely eliminated the for loop and replaced it with list comprehension logic. Definitely a drop in readability but a boost in performance over larger city counts.